Rename drivers to have a Driver instead of Loop suffix
