+++ test/Loop/EvLoopTest.php
@@ -2,16 +2,15 @@
 
 namespace Amp\Test\Loop;
 
-use Amp\Loop\EvLoop;
-use Amp\Test\LoopTest;
+use Amp\Loop\EvDriver;
 
 /**
  * @requires extension ev
  */
-class EvLoopLoopTest extends LoopTest {
+class EvDriverTest extends DriverTest {
     public function getFactory(): callable {
         return function () {
-            return new EvLoop;
+            return new EvDriver;
         };
     }
 }
